test: exercise production inline tool filter
