+++ documentation/lanes/preseed.md
@@ -636,7 +636,7 @@ export a default factory function. Pure helper modules such as
 factory alongside their named exports, or Pi aborts startup with
 `Extension does not export a valid factory function`.
 
-`inline-edit.ts` is the managed endpoint for [host-correlated native Pi editor results](../decisions/README.md#ad135-inline-chat-requires-one-host-correlated-result). It removes its result tool at session start, activates only that tool for `/codeflare-inline-edit`, dispatches through Pi's `ExtensionAPI.sendUserMessage`, isolates the current Inline turn suffix, and restores the exact previous tool list at settlement. The final initial-tool exposure filter preserves that exclusive result mode instead of replacing it with the normal panel tools. OpenAI Chat Completions and Responses payloads retain only the exact mandatory result function with parallel calls disabled. Panel turns therefore retain unrestricted tools and stored conversation without loading another Pi process. <!-- @impl: preseed/agents/pi/extensions/inline-edit.ts::registerInlineEditMode --> <!-- @impl: preseed/agents/pi/extensions/inline-edit.ts::constrainInlineOpenAiPayload --> <!-- @impl: preseed/agents/pi/extensions/capability.ts::registerInitialToolFilter --> <!-- @test: host/__tests__/pi-inline-edit-mode.test.js (REQ-IDE-025: inline result mode isolates provider tools and context then restores panel tools) -->
+`inline-edit.ts` is the managed endpoint for [host-correlated native Pi editor results](../decisions/README.md#ad135-inline-chat-requires-one-host-correlated-result). It removes its result tool at session start, activates only that tool for `/codeflare-inline-edit`, dispatches through Pi's `ExtensionAPI.sendUserMessage`, isolates the current Inline turn suffix, and restores the exact previous tool list at settlement. The final initial-tool exposure filter preserves that exclusive result mode instead of replacing it with the normal panel tools. OpenAI Chat Completions and Responses payloads retain only the exact mandatory result function with parallel calls disabled. Panel turns therefore retain unrestricted tools and stored conversation without loading another Pi process. <!-- @impl: preseed/agents/pi/extensions/inline-edit.ts::registerInlineEditMode --> <!-- @impl: preseed/agents/pi/extensions/inline-edit.ts::constrainInlineOpenAiPayload --> <!-- @impl: preseed/agents/pi/extensions/capability-helpers.ts::registerInitialToolFilter --> <!-- @test: host/__tests__/pi-inline-edit-mode.test.js (REQ-IDE-025: inline result mode isolates provider tools and context then restores panel tools) -->
 
 Native skill overrides include graphify
 ([REQ-AGENT-043](../../sdd/spec/agents.md#req-agent-043-graphify-build-mode-dispatch)
