+++ preseed/agents/pi/extensions/capability-helpers.ts
@@ -6,6 +6,17 @@ export type ToolActivationPi = {
   setActiveTools(names: string[]): void;
 };
 
+type SessionEntry = { type?: string; customType?: string; data?: unknown };
+type SessionContext = {
+  sessionManager?: {
+    getBranch?(): SessionEntry[];
+    getEntries?(): SessionEntry[];
+  };
+};
+type InitialToolFilterPi = ToolActivationPi & {
+  on(event: string, handler: (event: unknown, ctx: SessionContext) => void): void;
+};
+
 export type CapabilityMatch = {
   kind: "tool";
   name: string;
@@ -27,6 +38,17 @@ const CORE_TOOL_NAMES = [
 const TOOL_ACTIVATION_GROUPS: Readonly<Record<string, readonly string[]>> = {
   subagent: ["subagent", "get_subagent_result", "steer_subagent"],
 };
+const GOAL_STATE_ENTRY_TYPE = "goal-state";
+const INLINE_EDIT_RESULT_TOOL = "codeflare_submit_inline_result";
+const GOAL_TERMINAL_TOOLS = ["goal_complete", "goal_blocked"] as const;
+const UNFINISHED_GOAL_STATUSES = new Set([
+  "active",
+  "paused",
+  "blocked",
+  "usage_limited",
+  "budget_limited",
+  "queued",
+]);
 
 function unique(values: string[]): string[] {
   return [...new Set(values)];
@@ -39,6 +61,49 @@ export function initialActiveTools(pi: ToolActivationPi): string[] {
 
 export function isExclusiveActiveTool(activeTools: ReadonlySet<string>, toolName: string): boolean {
   return activeTools.size === 1 && activeTools.has(toolName);
+}
+
+function hasUnfinishedGoal(ctx: SessionContext): boolean {
+  let entries: SessionEntry[];
+  try {
+    entries = ctx.sessionManager?.getBranch?.() ?? ctx.sessionManager?.getEntries?.() ?? [];
+  } catch {
+    return false;
+  }
+  const latest = entries.filter((entry) => (
+    entry.type === "custom" && entry.customType === GOAL_STATE_ENTRY_TYPE
+  )).at(-1);
+  if (!latest || !latest.data || typeof latest.data !== "object") return false;
+  const goal = Reflect.get(latest.data, "goal");
+  if (!goal || typeof goal !== "object") return false;
+  const id = Reflect.get(goal, "id");
+  const status = Reflect.get(goal, "status");
+  return typeof id === "string"
+    && id.length > 0
+    && typeof status === "string"
+    && UNFINISHED_GOAL_STATUSES.has(status);
+}
+
+export function registerInitialToolFilter(pi: InitialToolFilterPi): void {
+  pi.on("before_agent_start", (_event, ctx) => {
+    const activeBeforeFilter = new Set(pi.getActiveTools());
+    // Inline Chat deliberately narrows the provider to one host-owned result tool.
+    // The final exposure filter runs later and must not replace that exclusive mode
+    // with the normal read/bash/edit/write/capability set.
+    if (isExclusiveActiveTool(activeBeforeFilter, INLINE_EDIT_RESULT_TOOL)) return;
+    const keepGoalTools = hasUnfinishedGoal(ctx)
+      || GOAL_TERMINAL_TOOLS.every((name) => activeBeforeFilter.has(name));
+    const initial = initialActiveTools(pi);
+    if (!keepGoalTools) {
+      pi.setActiveTools(initial);
+      return;
+    }
+    const registered = new Set(pi.getAllTools().map((tool) => tool.name));
+    pi.setActiveTools([
+      ...initial,
+      ...GOAL_TERMINAL_TOOLS.filter((name) => registered.has(name)),
+    ]);
+  });
 }
 
 export function activationGroup(name: string): string[] {
